feat: add `which` command
The `which` command will print the path to the active zig version. All
it really does is concatenate `ZVM_HOME` and the active version. If we
change how we store versions this will probably break, but it's good
enough for now.

If there is no active version, there will be no output and the script
will exit with a non-success code.